Use zipped csvs since Heroku doesn't support git-lfs
Since Heroku doesn't support git LFS, we originally explored downloading
 the file directly from the IRS website as part of the command. However,
the IRS blocks http requests from Heroku (our guess is that this is
to prevent bots/scrapers).

This implementation is not very elegant, but given we only update the
files occasionally, we can live with the ugliness and improve it in the
future if we need to.