Fix bugs in new pledge checkout flow for campaigns (#474)
* Fix bug in new pledge checkout flow for campaigns

The bug was that the oftw_wizard_controller was expecting currencies in upper
case but the view was providing them in lower case.

* Only handle verification successes for payments in ACSS debit

This logic-branch was seemingly being followed in the case of non-ACSS verifications triggering
the webhook. This raised (harmless) errors when 'UpdateCustomerAcssDebitDetails' tried to operate
on the payment methods assuming that they were of the ACSS debit type.

For more information on what the handle verification success method is for, see the PR that
introduced it, particularly the 4th bullet point of the description:
https://github.com/donational-org/donational-app/pull/396

* Trust html e.g. bold tags when rendering custom questions

These are always provided by trusted users (partners).