Fix custom donor questions bug for donation_box page
We had been silently marking some checkboxes as 'required' in html but
not displaying any information about this to the user, leaving them baffled
as to why they could not submit their pledge. Also, we had been marking
all checkboxes as required, heedless to whether they were in fact required.